Reposition declaration parsing ready to add tests
