+++ boa/src/syntax/parser/statement/declaration/hoistable/async_function_decl/tests.rs
@@ -1,0 +1,1 @@
+
