+++ boa/src/syntax/parser/statement/declaration/hoistable/function_decl/tests.rs
@@ -1,0 +1,1 @@
+
