+++ boa/src/syntax/parser/statement/declaration/hoistable/tests.rs
@@ -1,0 +1,1 @@
+
